fix(ci): resolve clippy lints and formatting in Newtype transition
- Fix empty line lints in keyforge-model newtypes
- Address truncation and unnecessary cast lints in SIMD engines
- Use idiomatic is_multiple_of in annealing reporter
- Verify full workspace parity with local lint and test pass

+++ libs/keyforge-physics/src/engines/intel_comet_lake.rs
@@ -135,7 +135,7 @@ unsafe fn score_layout_avx2(
         used,
     );
     for &code in pm.used_keys() {
-        let cand = pm.get(KeyCode(code as u16));
+        let cand = pm.get(KeyCode(code));
         if !cand.is_empty() {
             flat_map[code as usize] = cand[0];
         }
